requirements.txt
fastapi
starlette
python-dotenv
python-multipart
httpx
pandas
json5
pdfplumber
zipfile36
requests
uvicorn
